Management command optimization big DBs and use BigIntegerField

+++ onadata/apps/logger/migrations/0023_add_xform_attachment_storage_bytes.py
@@ -15,6 +15,6 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='xform',
             name='attachment_storage_bytes',
-            field=models.PositiveIntegerField(default=0),
+            field=models.BigIntegerField(default=0),
         ),
     ]
